#77 somewha works, now just need to change ordering of images
